Cria Header e armazena arquivos do @path
Nesta parte é feito o array pra armazenar nomes de arquivos lidos do comando @path e a verificação da extensão do parâmetro --config
